Move study-related files to dedicated packages
