+++ app/src/main/java/com/example/studyapp/data/study/Subtopic.kt
@@ -1,4 +1,4 @@
-package com.example.studyapp.data
+package com.example.studyapp.data.study
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
