+++ app/src/main/java/com/example/studyapp/data/study/TopicWithProgress.kt
@@ -1,4 +1,4 @@
-package com.example.studyapp.data
+package com.example.studyapp.data.study
 
 data class TopicWithProgress(
     val topic: Topic,
