+++ app/src/main/java/com/example/studyapp/data/study/TopicsRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package com.example.studyapp.data
+package com.example.studyapp.data.study
 
 import kotlinx.coroutines.flow.Flow
 import org.jetbrains.annotations.VisibleForTesting
