+++ app/src/main/java/com/example/studyapp/di/Modules.kt
@@ -3,12 +3,12 @@ package com.example.studyapp.di
 import android.content.Context
 import androidx.room.Room
 import com.example.studyapp.data.AppDatabase
-import com.example.studyapp.data.SubtopicDao
-import com.example.studyapp.data.SubtopicsRepository
-import com.example.studyapp.data.SubtopicsRepositoryImpl
-import com.example.studyapp.data.TopicDao
-import com.example.studyapp.data.TopicsRepository
-import com.example.studyapp.data.TopicsRepositoryImpl
+import com.example.studyapp.data.study.SubtopicDao
+import com.example.studyapp.data.study.SubtopicsRepository
+import com.example.studyapp.data.study.SubtopicsRepositoryImpl
+import com.example.studyapp.data.study.TopicDao
+import com.example.studyapp.data.study.TopicsRepository
+import com.example.studyapp.data.study.TopicsRepositoryImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
