+++ app/src/main/java/com/example/studyapp/ui/study/components/AdaptiveFAB.kt
@@ -1,4 +1,4 @@
-package com.example.studyapp.ui.components.study
+package com.example.studyapp.ui.study.components
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
