+++ app/src/main/java/com/example/studyapp/ui/study/components/PlaceholderColumn.kt
@@ -1,4 +1,4 @@
-package com.example.studyapp.ui.components
+package com.example.studyapp.ui.study.components
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
