+++ app/src/main/java/com/example/studyapp/ui/study/subtopics/SubtopicsViewModel.kt
@@ -1,14 +1,14 @@
-package com.example.studyapp.ui.subtopics
+package com.example.studyapp.ui.study.subtopics
 
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.studyapp.data.Subtopic
-import com.example.studyapp.data.SubtopicsRepository
-import com.example.studyapp.data.Topic
-import com.example.studyapp.data.TopicWithProgress
-import com.example.studyapp.data.TopicsRepository
-import com.example.studyapp.domain.GetTopicsWithProgressUseCase
+import com.example.studyapp.data.study.Subtopic
+import com.example.studyapp.data.study.SubtopicsRepository
+import com.example.studyapp.data.study.Topic
+import com.example.studyapp.data.study.TopicWithProgress
+import com.example.studyapp.data.study.TopicsRepository
+import com.example.studyapp.domain.study.GetTopicsWithProgressUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
