+++ app/src/main/java/com/example/studyapp/ui/study/topics/TopicsScreen.kt
@@ -1,4 +1,4 @@
-package com.example.studyapp.ui.topics
+package com.example.studyapp.ui.study.topics
 
 import androidx.annotation.VisibleForTesting
 import androidx.compose.animation.AnimatedContent
@@ -26,15 +26,15 @@ import androidx.compose.ui.tooling.preview.PreviewScreenSizes
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.example.studyapp.R
-import com.example.studyapp.data.Topic
-import com.example.studyapp.data.TopicWithProgress
-import com.example.studyapp.ui.components.DockedSearchBar
-import com.example.studyapp.ui.components.PlaceholderColumn
-import com.example.studyapp.ui.components.SearchAppBar
-import com.example.studyapp.ui.components.study.AdaptiveFAB
-import com.example.studyapp.ui.components.study.LoadingIndicatorBox
-import com.example.studyapp.ui.components.study.SaveTopicDialog
-import com.example.studyapp.ui.components.study.TopicsLazyColumn
+import com.example.studyapp.data.study.Topic
+import com.example.studyapp.data.study.TopicWithProgress
+import com.example.studyapp.ui.study.components.AdaptiveFAB
+import com.example.studyapp.ui.study.components.DockedSearchBar
+import com.example.studyapp.ui.study.components.LoadingIndicatorBox
+import com.example.studyapp.ui.study.components.PlaceholderColumn
+import com.example.studyapp.ui.study.components.SaveTopicDialog
+import com.example.studyapp.ui.study.components.SearchAppBar
+import com.example.studyapp.ui.study.components.TopicsLazyColumn
 import com.example.studyapp.ui.theme.StudyAppTheme
 
 @Composable
