+++ app/src/test/java/com/example/studyapp/SubtopicViewModelTest.kt
@@ -1,9 +1,9 @@
 package com.example.studyapp
 
 import androidx.lifecycle.SavedStateHandle
-import com.example.studyapp.data.Subtopic
-import com.example.studyapp.data.SubtopicsRepository
-import com.example.studyapp.ui.subtopic.SubtopicViewModel
+import com.example.studyapp.data.study.Subtopic
+import com.example.studyapp.data.study.SubtopicsRepository
+import com.example.studyapp.ui.study.subtopic.SubtopicViewModel
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.every
