+++ app/src/test/java/com/example/studyapp/TopicsViewModelTest.kt
@@ -1,10 +1,10 @@
 package com.example.studyapp
 
-import com.example.studyapp.data.Topic
-import com.example.studyapp.data.TopicWithProgress
-import com.example.studyapp.data.TopicsRepository
-import com.example.studyapp.domain.GetTopicsWithProgressUseCase
-import com.example.studyapp.ui.topics.TopicsViewModel
+import com.example.studyapp.data.study.Topic
+import com.example.studyapp.data.study.TopicWithProgress
+import com.example.studyapp.data.study.TopicsRepository
+import com.example.studyapp.domain.study.GetTopicsWithProgressUseCase
+import com.example.studyapp.ui.study.topics.TopicsViewModel
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.impl.annotations.MockK
