+++ gradle/libs.versions.toml
@@ -2,8 +2,8 @@
 agp = "8.11.0"
 coilCompose = "2.7.0"
 coreSplashscreen = "1.0.1"
-junitJupiterApi = "5.13.2"
-junitJupiterEngine = "5.13.2"
+junitJupiterApi = "5.13.3"
+junitJupiterEngine = "5.13.3"
 kotlin = "2.2.0"
 coreKtx = "1.16.0"
 kotlinxCoroutinesTest = "1.10.2"
@@ -21,6 +21,9 @@ serialization = "1.9.0"
 glance = "1.1.1"
 uiTestJunit4 = "1.8.3"
 uiTestManifest = "1.8.3"
+androidx-credentials = "1.5.0"
+androidx-credentials-play-services-auth = "1.5.0"
+uiTooling = "1.8.3"
 
 [libraries]
 androidx-adaptive = { module = "androidx.compose.material3.adaptive:adaptive" }
@@ -32,6 +35,8 @@ androidx-material3-adaptive-navigation-suite = { module = "androidx.compose.mate
 androidx-ui-test-junit4 = { module = "androidx.compose.ui:ui-test-junit4", version.ref = "uiTestJunit4" }
 androidx-ui-test-manifest = { module = "androidx.compose.ui:ui-test-manifest", version.ref = "uiTestManifest" }
 coil-compose = { module = "io.coil-kt:coil-compose", version.ref = "coilCompose" }
+androidx-credentials = { module = "androidx.credentials:credentials", version.ref = "androidx-credentials" }
+androidx-credentials-play-services-auth = { module = "androidx.credentials:credentials-play-services-auth", version.ref = "androidx-credentials-play-services-auth" }
 
 androidx-lifecycle-runtime-ktx = { group = "androidx.lifecycle", name = "lifecycle-runtime-ktx", version.ref = "lifecycleRuntimeKtx" }
 androidx-activity-compose = { group = "androidx.activity", name = "activity-compose", version.ref = "activityCompose" }
@@ -54,6 +59,7 @@ navigation-compose = { module = "androidx.navigation:navigation-compose", versio
 kotlinx-serialization-json = { module = "org.jetbrains.kotlinx:kotlinx-serialization-json", version.ref = "serialization" }
 glance-appwidget = { group = "androidx.glance", name = "glance-appwidget", version.ref = "glance" }
 glance-material = { group = "androidx.glance", name = "glance-material3", version.ref = "glance" }
+androidx-ui-tooling = { group = "androidx.compose.ui", name = "ui-tooling", version.ref = "uiTooling" }
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
